moving controls into its own component
